LayoutTests/inspector/elements-panel-rewrite-href.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
<html>
<head>
<link rel="stylesheet" src="/bogusSheet1.css">
<link rel="stylesheet" src="bogusSheet2.css">
<script src="../http/tests/inspector/inspector-test.js"></script>
<script src="elements-tests.js"></script>
<script>

function doit()
{
    function callback(result)
    {
        for (var i = 0; i < result.length; ++i)
            output("[" + result[i] + "]");
        notifyDone();
    }
    evaluateInWebInspector("frontend_testRewriteHrefs", callback);
}


// Frontend functions.

function frontend_testRewriteHrefs(testController)
{
    testController.waitUntilDone();
    // Need test to be async to expand whole the tree first.
    frontend_expandDOMSubtreeAndRun(testController, null, frontend_testRewriteHrefsContinuation);
}

function frontend_testRewriteHrefsContinuation(testController)
{
    var innerMapping = WebInspector.domAgent._idToDOMNode;
    var result = [];

    for (var nodeId in innerMapping) {
        var node = innerMapping[nodeId];
        if (node.nodeName === "LINK" || node.nodeName === "SCRIPT") {
            var segments = [];
            var href = WebInspector.resourceURLForRelatedNode(node, node.getAttribute("src") || node.getAttribute("href"));
            if (!href) {
                segments.push("<empty>");
                continue;
            }
            var match = href.match(WebInspector.URLRegExp);
            if (!match)
                testController.notifyDone("FAIL: no URL match for <" + href + ">");
            var split = match[4].split("/");
            for (var i = split.length - 1, j = 0; j < 3 && i >= 0; --i, ++j)
                segments.push(split[i]);
            result.push(segments.reverse());
        }
    }
    testController.notifyDone(result);
}

</script>
</head>

<body onload="onload()">
<p>
Tests that src and href element targets are rewritten properly.
</p>

</body>
</html>
