LayoutTests/inspector/elements-panel-rewrite-href-expected.txt

Tests that src and href element targets are rewritten properly.

[,bogusSheet1.css]
[LayoutTests,inspector,bogusSheet2.css]
[tests,inspector,inspector-test.js]
[LayoutTests,inspector,elements-tests.js]
